+++ tests/test_sdk_save.py
@@ -2,6 +2,9 @@ import unittest
 
 from keras.applications import vgg19
 
+from aim.sdk.save import Checkpoint
+from aim.sdk.save.save import save
+
 
 class TestSDKSave(unittest.TestCase):
     def test_sdk_save(self):
@@ -12,5 +15,7 @@ class TestSDKSave(unittest.TestCase):
                                 pooling=None,
                                 classes=1000)
 
-        # TODO: test sdk save
-        # vgg_model.save('path_to_my_model.h5')
+        for e in range(3):
+            save(Checkpoint('vgg_model',
+                            'vgg_checkpoint_{}'.format(e),
+                            vgg_model, epoch=e, meta={'foo': 'bar'}))